requirements.txt
PyQt5
Pillow
pyautogui
requests
